backend/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Security-Too==5.0.2
Flask-Caching==2.0.1
celery[redis]
requests==2.28.1
weasyprint==57.2